Refactoring admin server app creation
