+++ lib/core/mount.js
@@ -93,7 +93,7 @@ function mount(events) {
 	// unless the headless option is set (which disables the Admin UI),
 	// add the Admin UI's Static Router for public resources
 	if (!this.get('headless')) {
-		app.use('/keystone', require('../../admin/server/app/static'));
+		app.use('/keystone', require('../../admin/server').createStaticRouter(this));
 	}
 
 	require('../../server/bindLessMiddleware')(this, app);
